+++ packages/experimental/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- [08-17-19] Party properties (via special item).
 - [08-17-19] PartyManager real-time update and coldstart.
 - [08-17-20] Pipeline with feed-store-iterator (using party-processor).
 - [08-17-20] Pipeline with basic party-procesor.
@@ -27,7 +28,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Backlog
 
 - Item models (properties and data).
-- Party properties (via special item).
 - Ensure streams are closed when objects are destroyed (on purpose or on error); error handling. Asserts vs Errors?
 - Reactive components (Database, Party, Item, Model) with event propagation.
 - Event handlers: global state to warn of leaks when system shuts down (show graph).